requirements.txt
# Network and Security
scapy>=2.5.0
netaddr>=0.8.0
netifaces>=0.11.0
python-nmap>=0.7.1
cryptography>=41.0.1
pyOpenSSL>=23.2.0
paramiko>=3.3.1
dnspython>=2.4.2
pycryptodome>=3.18.0

# Database
mysqlclient>=2.2.0
psycopg2-binary>=2.9.7
pymongo>=4.4.1
redis>=4.6.0

# HTTP and Web
requests>=2.31.0
aiohttp>=3.8.5
beautifulsoup4>=4.12.2
lxml>=4.9.3
httpx>=0.24.1

# Async and Concurrency
asyncio>=3.4.3
aiofiles>=23.1.0
gevent>=23.7.0

# Data Processing
numpy>=1.25.2
pandas>=2.0.3
pyarrow>=13.0.0

# System Management
psutil>=5.9.5
py-cpuinfo>=9.0.0

# Logging
rich>=13.5.2
loguru>=0.7.0
structlog>=23.1.0

# File Processing
python-magic>=0.4.27
pyyaml>=6.0.1

# CLI
typer>=0.9.0
click>=8.1.7
prompt-toolkit>=3.0.39
tqdm>=4.65.0

# Performance
cython>=3.0.2
ujson>=5.8.0
msgpack>=1.0.5

# Support Libraries
six>=1.16.0
python-dateutil>=2.8.2
pytz>=2023.3
urllib3>=2.0.4

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
mypy>=1.5.1
coverage>=7.3.0
pylint>=2.17.5
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
autopep8>=2.0.2

# Error Tracking
sentry-sdk>=1.29.2

# Documentation
sphinx>=7.1.2
mkdocs>=1.5.2

# Package Management
pip>=23.2.1
setuptools>=68.1.2
wheel>=0.41.1